+++ packages/file-input-react/src/index.ts
@@ -1,0 +1,6 @@
+export type { FileProps } from "./File";
+export { File } from "./File";
+export type { FileInputProps } from "./FileInput";
+export { FileInput } from "./FileInput";
+export type { FileInputFile, FileInputFileValidation, FileInputFileState } from "./types";
+export { upload } from "./utils";