+++ portal/gatsby-browser.tsx
@@ -18,6 +18,7 @@ import "@fremtind/jkl-checkbox/checkbox.min.css";
 import "@fremtind/jkl-datepicker/datepicker.min.css";
 import "@fremtind/jkl-description-list/description-list.min.css";
 import "@fremtind/jkl-feedback/feedback.min.css";
+import "@fremtind/jkl-file-input/file-input.min.css";
 import "@fremtind/jkl-hamburger/hamburger.min.css";
 import "@fremtind/jkl-icon-button/icon-button.min.css";
 import "@fremtind/jkl-input-group/input-group.min.css";
@@ -41,6 +42,7 @@ import "@fremtind/jkl-cookie-consent/cookie-consent.min.css";
 import "@fremtind/jkl-contextual-menu/contextual-menu.min.css";
 import "@fremtind/jkl-expand-button/expand-button.min.css";
 import "@fremtind/jkl-footer/footer.min.css";
+import "@fremtind/jkl-progress-bar/progress-bar.min.css";
 import "./src/components/Typography/typography.scss";
 
 initTabListener();